add splash
