+++ app/src/main/java/com/jys/catsapp/core/MainActivity.kt
@@ -3,7 +3,6 @@ package com.jys.catsapp.core
 import android.app.Activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import androidx.activity.SystemBarStyle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -13,14 +12,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.jys.catsapp.presentation.HomeScreen
 import com.jys.catsapp.core.theme.CatsAPPTheme
 import com.jys.catsapp.navigation.Navigation
-import org.koin.androidx.compose.KoinAndroidContext
-import org.koin.core.annotation.KoinExperimentalAPI
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -34,8 +27,6 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
-
 
 
 @Composable
