+++ app/src/main/java/com/jys/catsapp/core/common/Constants.kt
@@ -1,0 +1,5 @@
+package com.jys.catsapp.core.common
+
+object Constants {
+    const val VIDEO_SAMURAI_CAT = "asset:///samurai_cat.mp4"
+}
