+++ app/src/main/java/com/jys/catsapp/core/common/DimensUtil.kt
@@ -1,4 +1,15 @@
 package com.jys.catsapp.core.common
 
-class DimensUtil {
+import androidx.compose.ui.unit.dp
+
+object DimensUtil {
+    val Dimens1Dp = 1.dp
+    val Dimens2Dp = 2.dp
+    val Dimens4Dp = 4.dp
+    val Dimens8Dp = 8.dp
+    val Dimens10Dp = 10.dp
+    val Dimens12Dp = 12.dp
+    val Dimens16Dp = 16.dp
+    val Dimens32Dp = 32.dp
+    val Dimens150Dp = 150.dp
 }
