+++ app/src/main/java/com/jys/catsapp/data/localDB/PhotoEntity.kt
@@ -19,7 +19,7 @@ data class PhotoEntity(
     val photographerUrl: String?,
     val width: Int?,
     val pageNumber: Int? = 0,
-    var position: Int? = 0 // Campo agregado para el orden
+    var position: Int? = 0
 )
 
 
@@ -34,32 +34,4 @@ data class SrcEntity(
     val tiny: String?
 )
 
-fun Photo.toEntity(page: Int): PhotoEntity {
-    return PhotoEntity(
-        id = this.id,
-        width = this.width,
-        height = this.height,
-        url = this.url,
-        photographer = this.photographer,
-        photographerUrl = this.photographerUrl,
-        photographerId = this.photographerId,
-        avgColor = this.avgColor ?: "",
-        src = this.src.toEntity(),
-        alt = this.alt,
-        liked = this.liked,
-        pageNumber = page
-    )
-}
 
-fun Src.toEntity(): SrcEntity {
-    return SrcEntity(
-        original = this.original,
-        large2x = this.large2x,
-        large = this.large,
-        medium = this.medium,
-        small = this.small,
-        portrait = this.portrait,
-        landscape = this.landscape,
-        tiny = this.tiny
-    )
-}
