+++ app/src/main/java/com/jys/catsapp/di/HomeModule.kt
@@ -9,6 +9,7 @@ import com.jys.catsapp.presentation.HomeViewModel
 
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.core.module.dsl.factoryOf
+import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
