+++ app/src/main/java/com/jys/catsapp/navigation/Navigation.kt
@@ -4,26 +4,32 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.jys.catsapp.core.common.Constants.VIDEO_SAMURAI_CAT
+import com.jys.catsapp.core.common.VideoBackgroundScreen
+import com.jys.catsapp.navigation.NavigationRoute.DASHBOARD
+import com.jys.catsapp.navigation.NavigationRoute.HOME_WITHOUT_ROOM
+import com.jys.catsapp.navigation.NavigationRoute.HOME_WITH_ROOM
 import com.jys.catsapp.presentation.HomeScreen
 import com.jys.catsapp.presentation.HomeScreenWithRoom
 import com.jys.catsapp.presentation.components.DashBoard
-import com.jys.catsapp.presentation.components.VideoBackgroundScreen
 
 @Composable
 fun Navigation() {
     val navController = rememberNavController()
     NavHost(
         navController = navController,
-        startDestination = "home",
+        startDestination = DASHBOARD,
     ) {
-        composable("home") {
-            VideoBackgroundScreen { DashBoard(navController = navController) }
+        composable(DASHBOARD) {
+            VideoBackgroundScreen(
+                pathVideo = VIDEO_SAMURAI_CAT,
+                content = { DashBoard(navController = navController) }
+            )
         }
-        composable("homeWithOutRoom") {
+        composable(HOME_WITHOUT_ROOM) {
             HomeScreen()
         }
-
-        composable("homeWithRoom") {
+        composable(HOME_WITH_ROOM) {
             HomeScreenWithRoom()
         }
     }
