+++ app/src/main/java/com/jys/catsapp/navigation/NavigationRoute.kt
@@ -1,0 +1,7 @@
+package com.jys.catsapp.navigation
+
+object NavigationRoute {
+    const val DASHBOARD = "DashBoard"
+    const val HOME_WITHOUT_ROOM = "homeWithOutRoom"
+    const val HOME_WITH_ROOM = "homeWithRoom"
+}
