+++ app/src/main/res/values/themes.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.CatsAPP" parent="android:Theme.Material.Light.NoActionBar" />
+    <style name="Theme.CatsAPP" parent="android:Theme.Material.Light.NoActionBar" >
+        <item name="android:windowIsTranslucent">true</item>
+    </style>
+
 </resources>
+
